Inject settings on load
Added first dummy controls
